backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
groq
gunicorn
PyJWT
bcrypt